+++ airflow/models/dagparam.py
@@ -15,12 +15,7 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from typing import Any, Dict, Callable
-
-from inspect import signature
-
-
-import functools
+from typing import Any, Dict
 
 
 class DagParam:
@@ -55,31 +50,3 @@ class DagParam:
         in respectable context.
         """
         return context.get('params', {}).get(self._name, self._default)
-
-
-def dag(*dag_args, **dag_kwargs):
-    """
-    Python dag decorator. Wraps a function into an Airflow DAG.
-    Accepts kwargs for operator kwarg. Can be used to parametrize DAGs.
-
-    :param dag_args: Arguments for DAG object
-    :type dag_args: list
-    :param dag_kwargs: Kwargs for DAG object.
-    :type dag_kwargs: dict
-    """
-    def wrapper(f: Callable):
-        from airflow.models.dag import DAG
-        dag_sig = signature(DAG.__init__)
-        dag_sig = dag_sig.bind_partial(*dag_args, **dag_kwargs)
-        @functools.wraps(f)
-        def factory(*args, **kwargs):
-            f_sig = signature(f).bind(*args, **kwargs)
-            f_sig.apply_defaults()
-            with DAG(*dag_sig.args, dag_id=f.__name__, **dag_sig.kwargs) as dag_obj:
-                f_kwargs = {}
-                for name, value in f_sig.arguments.items():
-                    f_kwargs[name] = dag_obj.param(name, value)
-                f(**f_kwargs)
-            return dag_obj
-        return factory
-    return wrapper
